requirements.txt
aiohttp>=3.9.0
gspread>=5.12.0
google-auth>=2.23.0
httpx>=0.25.0
python-dotenv>=1.0.0